BoltHttp_football-relay/BoltHttp_football-relay/requirements.txt
nats-py==2.2.0
aiohttp==3.11.9